+++ public/js/editor/delete.mjs
@@ -7,18 +7,35 @@ if (deleteButton) {
 	deleteButton.addEventListener('click', async () => {
 		const noteId = JSON.parse(document.getElementById('metadata--note-id').textContent);
 
-		const deleteConfirmation = await dialog.confirm({
+		const data = await dialog.form({
 			title: 'Delete note',
-			message: 'Are you sure you want to delete this comment? This action is permanent.',
+			message: 'Type <code>DELETE</code> to confirm. This action is permanent.',
 			cancelIcon: 'close',
 			cancelText: 'Cancel',
 			confirmIcon: 'delete',
 			confirmText: 'Delete forever',
-			destructive: true
+			destructive: true,
+			fields: [
+				{
+					name: 'inputText',
+					label: 'Delete',
+					required: true,
+					type: 'text',
+					icon: 'approval',
+				},
+			],
 		});
 
-		if (!deleteConfirmation)
+		if (data === null)
 			return;
+
+		if (data.inputText !== 'DELETE') {
+			toast({
+				title: 'Canceled',
+				message: `Note deletion canceled, you didn't type <code>DELETE</code>`
+			});
+			return;
+		}
 
 		try {
 			const response = await fetch(`/api/notes/${noteId}`, {